requirements.txt
algoliasearch==1.14.0
asn1crypto==0.23.0
awscli==1.11.178
boto3==1.4.7
botocore==1.7.36
certifi==2017.7.27.1
cffi==1.11.2
chardet==3.0.4
click==6.7
colorama==0.3.7
cryptography==2.1.2
cycler==0.10.0
decorator==4.1.2
dlib==19.7.99
docutils==0.14
face-recognition==1.0.0
face-recognition-models==0.3.0
gcloud==0.17.0
googleapis-common-protos==1.5.3
httplib2==0.10.3
idna==2.6
jmespath==0.9.3
jws==0.1.3
matplotlib==2.1.0
networkx==2.0
numpy==1.13.3
oauth2client==3.0.0
olefile==0.44
Pillow==4.3.0
protobuf==3.4.0
pyasn1==0.3.7
pyasn1-modules==0.1.4
pycparser==2.18
pycryptodome==3.4.3
pyfcm==1.4.2
pyOpenSSL==17.3.0
pyparsing==2.2.0
Pyrebase==3.0.27
python-dateutil==2.6.1
python-jwt==2.0.1
pytz==2017.2
PyWavelets==0.5.2
PyYAML==3.12
requests==2.18.4
requests-toolbelt==0.7.0
rsa==3.4.2
s3transfer==0.1.11
scikit-image==0.13.1
scipy==0.19.1
six==1.11.0
urllib3==1.22
